Human and mouse Plots
These are plots from the human and mouse data we worked with, they're png images because the actual pdfs were too large to upload, even zipped. The names 1e01 and 5e05 refer to a length limit we put on the x axis to better view the data. Let me know if anyone has any questions!...sorry this is so late....